+++ src/main/java/racing/util/NumberUtil.java
@@ -1,4 +1,4 @@
-package racing;
+package racing.util;
 
 import java.util.Random;
