+++ src/test/java/racing/domain/CarTest.java
@@ -1,7 +1,8 @@
-package racing;
+package racing.domain;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import racing.domain.Car;
 
 import static org.assertj.core.api.Assertions.assertThat;
