+++ src/test/java/racing/domain/enums/MoveStatusTest.java
@@ -1,7 +1,8 @@
-package racing;
+package racing.domain.enums;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import racing.domain.enums.MoveStatus;
 
 import static org.assertj.core.api.Assertions.assertThat;
