+++ src/test/java/racing/util/NumberUtilTest.java
@@ -1,7 +1,8 @@
-package racing;
+package racing.util;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import racing.util.NumberUtil;
 
 import static org.assertj.core.api.Assertions.assertThat;
